requirements.txt
djongo == 1.2.36
djangorestframework
markdown
django-filter 